sua rate

+++ Diamond.Entities/Model/DiamondModel.cs
@@ -19,7 +19,7 @@ namespace Diamond.Entities.Model
         public string Cut { get; set; }
         public decimal DiameterMM { get; set; }
         public string Color { get; set; }
-        public int MarkupRate { get; set; }
+        public decimal MarkupRate { get; set; }
         public int MarkupPrice { get; set; }
         public string Description { get; set; }
         public int BasePrice { get; set; }
